refactor : gamer와 관련없는 클래스 패키지 이동
